Closes #107, added JS Unit tests.
Created a separate client folder to hold all clients that we intend to
implement outside of Java for MatsSocket, and added the javascript
client there. The folder has been structured as a node module, so that
it can be published to npm.

To facilitate modules inclusion in both node and broswer environments,
the MatsSocket.js file was enclosed in a UMD module decleration. This
also has the benefit of working nicely with other frameworks, such as
require.js and Webpack. It does change MatsSocket from being a root
level constructor, to be namespaced under mats for browsers.

Created gradle tasks to start a test Mats WebSocket server, that can
then be used as a backend for integration tests. Also created some
sample integration tests based on the existing html/javascript code in
the project. Currently these output only to the console, so some further
work will be needed to expose the test results in reporting.

The sample html files where also moved into the javascript client, and
have their own serve task from package.json that can either be invoked
from the gradle task (serveHtml) or via yarn/npm run serve. This is
because the js module and the sample html files are no longer in the
same path.

I contribute this material in accordance with the signed SCA.
