+++ mats-websockets/client/javascript/pub/static/index.html
