+++ mats-websockets/client/javascript/pub/static/simple_tests.html
@@ -6,7 +6,7 @@
 <script src="MatsSocket.js"></script>
 
 <script>
-    var matsSocket = new MatsSocket("TestApp", "1.2.3",
+    var matsSocket = new mats.MatsSocket("TestApp", "1.2.3",
         ["ws://localhost:8080/matssocket/json", "ws://localhost:8081/matssocket/json"]);
     matsSocket.logging = true;
